requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
cryptography>=42.0.0
httpx>=0.27.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
redis>=5.0.0
